requirements.txt
click==8.0.4
dataclasses==0.8
distlib==0.3.6
filelock==3.4.1
Flask==2.0.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
platformdirs==2.4.0
PyMySQL==1.0.2
python-dotenv==0.20.0
six==1.16.0
typing-extensions==4.1.1
virtualenv==20.17.1
Werkzeug==2.0.3
zipp==3.6.0
